refactor: reorder page flow
